python/kafka-project/pyproject.toml
[project]
authors = [
    { name = "Stefano Babini", email = "stefano.babini5@studio.unibo.it" },
]
name = "kafka-project"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "confluent-kafka>=2.8.0,<3.0.0",
    "psycopg2-binary>=2.9.10,<3.0.0",
]
[tool.poetry]
packages = [{ include = "src" }]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
